AMBARI-6432 Add FreeIPA (Redhat IDM) support to Ambari
This patch adds support for FreeIPA to Ambari. It depends on the
ipa-admin tools being installed on the ambari host. For the principals
it depends on a suitable password policy or being able to write to
krbPasswordExpiry.

It can be enabled by turning on enableIpa on the experimental features page.

Note:
* IPA requires user principal names to be in lowercase hence a toLower Function has been added that converts the clustername to lowercase. This is applied to all templates.